Amélioration de l'API et réécriture de la page schedule

+++ Application/parkaps-server/app/Http/Controllers/AvailabilityController.php
@@ -8,6 +8,7 @@
 
 namespace App\Http\Controllers;
 
+use App\DailyAvailability;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use App\Availability;
@@ -20,24 +21,23 @@ class AvailabilityController
         'start.numeric' => 'La date de début doit être un entier au format timestamp',
         'end.required' => "Une date de fin est requise",
         'end.numeric' => 'La date de fin doit être un entier au format timestamp',
-        'daily.boolean' => "Le paramètre \"jounalié\" doit être un booléen",
+        'daily.required' => "Le paramètre \" journalier\" est requis",
+        'daily.boolean' => "Le paramètre \"jounalier\" doit être un booléen",
         'carParkId.required' => "Un identifiant d'une place de parking est requise",
         'carParkId.numeric' => "L'identifiant de la place de parking doit être un entier positif",
     ];
 
 
 
-    public function search(Request $request)
+    public function get($id)
     {
-        $request->validate([
-            'carParkId' => 'required|numeric',
-        ], $this->messages);
-
-        $availabilites = DB::table('availabilities')->get()->where('car_park_id', $request->carParkId);
+        $availabilities = DB::table('availabilities')->where('car_park_id', $id)->where('end', '>=', Carbon::now()->subMonth(1)->toDateTimeString())->get();
+        $dailyAvailabilities = DB::table('daily_availabilities')->where('car_park_id', $id)->where('end', '>=', Carbon::now()->subMonth(1)->toDateTimeString())->get();
 
         return response()->json([
-            'availabilites' => $availabilites,
-        ], 201);
+            'availabilities' => $availabilities,
+            'daily_availabilities' => $dailyAvailabilities
+        ], 200);
 
     }
 
@@ -57,59 +57,56 @@ class AvailabilityController
            'carParkId' => 'required|numeric'
        ], $this->messages);
 
-       $availability = new Availability([
-           'start' => Carbon::createFromTimestampMs($request->start),
-           'end' =>  Carbon::createFromTimestampMs($request->end),
-           'daily' => $request->daily,
-           'car_park_id' => $request->carParkId
-       ]);
+       $availability = null;
+       if($request->daily){
+           $availability = new DailyAvailability([
+               'start' => Carbon::createFromTimestampMs($request->start),
+               'end' =>  Carbon::createFromTimestampMs($request->end),
+               'car_park_id' => $request->carParkId
+           ]);
+       } else {
+           $availability = new Availability([
+               'start' => Carbon::createFromTimestampMs($request->start),
+               'end' =>  Carbon::createFromTimestampMs($request->end),
+               'car_park_id' => $request->carParkId
+           ]);
+       }
        $availability->save();
 
         return response()->json([
-            'created' => $availability,
+            'availability' => $availability,
         ], 201);
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param Availability $availability
-     * @return void
-     */
-    public function show(Availability $availability)
-    {
-        //
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request $request
-     * @param Availability $availability
-     * @return void
-     */
-    public function update(Request $request, Availability $availability)
-    {
-        //
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param Availability $availability
+     * @param $id
+     * @param Request $request
      * @return \Illuminate\Http\Response
-     * @throws \Exception
      */
-    public function destroy(Request $request)
+    public function delete($id, Request $request)
     {
         $request->validate([
-            'id' => 'required|numeric'
+            'daily' => 'required|boolean'
         ]);
-        $availability = Availability::find($request->id);
-        $availability->delete();
 
-        return response()->json([
-            'deleted' => $availability,
-        ], 201);
+        if($request->daily){
+            $availability = DailyAvailability::find($id);
+            if($availability->id == $request->user()->id){
+                $availability->delete();
+            } else {
+                return response('Forbidden', 403);
+            }
+        } else {
+            $availability = Availability::find($id);
+            if($availability->id == $request->user()->id){
+                $availability->delete();
+            } else {
+                return response('Forbidden', 403);
+            }
+        }
+
+        return response()->json('Deleted',200);
     }
 }
